requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.14.0,<3.15
django-health-check>=3.17.0,<3.18
gunicorn>=20.1.0,<20.2
requests==2.28.2
psycopg2>=2.9.5,<2.10
